requirements.txt
atpublic==0.5
attrs==17.4.0
backcall==0.1.0
bleach==3.1.0
Cerealizer==0.8.3
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
fenics==2019.1.0
fenics-dijitso==2019.1.0
fenics-dolfin==2019.1.0
fenics-ffc==2019.1.0.post0
fenics-fiat==2019.1.0
fenics-ufl==2019.1.0
flufl.lock==3.2
imageio==2.15.0
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.2
Jinja2==2.10.3
json5==0.8.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyterlab==1.2.4
jupyterlab-server==1.0.6
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==8.0.2
mpi4py==3.0.3
mpmath==1.1.0
mshr==2019.1.0
nbconvert==5.6.1
nbformat==5.0.3
networkx==2.5.1
notebook==6.0.2
numpy==1.19.5
olefile==0.45.1
packaging==21.3
pandocfilters==1.4.2
parso==0.5.2
petsc4py==3.12.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.4.0
pkgconfig==1.5.1
pluggy==0.6.0
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
py==1.5.2
pybind11==2.2.4
pyDeprecate==0.3.2
Pygments==2.5.2
PyGObject==3.26.1
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==3.3.2
python-apt==1.6.4
python-dateutil==2.8.1
PyWavelets==1.1.1
pyzmq==18.1.1
qtconsole==4.6.0
scikit-image==0.17.2
scipy==1.5.4
Send2Trash==1.5.0
six==1.11.0
slepc4py==3.12.0
Soya3==0.1
sympy==1.1.1
terminado==0.8.3
testpath==0.4.4
tifffile==2020.9.3
torch>=2.0.0
torchvision>=0.15.0
torchmetrics==0.8.2

# Additional requirements for new framework
pyyaml>=6.0
tqdm>=4.65.0
timm>=0.9.0
einops>=0.6.0
tornado==6.0.3
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.22
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==0.6.0
